move Rep to shared/Rep
